Additional Work
Not sure how many commits I've done but I updated tutorial.html and moved the image from the top to the bottom, I did the same thing with the image on index.html as well. The final thing I did was work more on game.html and I made it even better.