More Updates
I've done some more updates. I spend over an hour trying to figure out how to create a custom alert which pops up when you click on the buttons, but I couldn't figure it out. I got one of the alerts working but I couldn't edit the styling to how I wanted to so I abandoned it.

I've put this here for reference for the future.
https://www.w3schools.com/howto/howto_js_alert.asp

It almost worked but it never worked :(